LPS-61860 rename packages
